Move source files into src directory
